+++ server/src/modules/testPaper/dto/request/GradeTestPaperQuestionRequest.ts
@@ -7,10 +7,6 @@ class GradeTestPaperQuestionRequest {
   @IsInt()
   public testPaperQuestionId: number;
 
-  @ApiProperty()
-  @IsInt()
-  public questionId: number;
-
   @ApiProperty({
     enum: QuestionType,
     enumName: 'Question Type',
